First of MapLibre pushes to dev.  Closes #3148 and includes most of #3143
