+++ appv2/src/UI/Map2/map.css
@@ -1,0 +1,7 @@
+.userTrackingMarker {
+    display: block;
+        border: none;
+        border-radius: 50%;
+        cursor: pointer;
+        padding: 0;
+}
